// src/app/pages/services/data-access/api.enum.ts
export const enum EndPointUrl {
    SERVICES_CATEGORIES_ALL = "services/categories/all",
    SERVICES_CATEGORIES_CREATE = "services/categories/create",
    SERVICES_CATEGORIES_UPDATE = "services/categories/update",
    SERVICES_CATEGORIES_DELETE = "services/categories/delete",
    SERVICES_ALL = "services/all",
    SERVICES_ADD = "services/add",
    SERVICES_UPDATE = "services/update",
    SERVICES_DELETE = "services/delete",
    SERVICES_ROLES_ALL = "services/roles/all",
}